Moved window corner settings into a separate class to use everywhere.
