+++ NewsMaking.DesktopApp/Misc/DWMWINDOWATTRIBUTE.cs
@@ -1,0 +1,9 @@
+﻿namespace NewsMaking.DesktopApp.Misc
+{
+				// The enum flag for DwmSetWindowAttribute's second parameter, which tells the function what attribute to set.
+				// Copied from dwmapi.h
+				public enum DWMWINDOWATTRIBUTE
+				{
+								DWMWA_WINDOW_CORNER_PREFERENCE = 33
+				}
+}
