+++ NewsMaking.DesktopApp/Views/MainDisplayView.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using NewsMaking.DesktopApp.Misc;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.InteropServices;
@@ -16,42 +17,16 @@ using System.Windows.Shapes;
 
 namespace NewsMaking.DesktopApp.Views
 {
-				// The enum flag for DwmSetWindowAttribute's second parameter, which tells the function what attribute to set.
-				// Copied from dwmapi.h
-				public enum DWMWINDOWATTRIBUTE
-				{
-								DWMWA_WINDOW_CORNER_PREFERENCE = 33
-				}
-
-				// The DWM_WINDOW_CORNER_PREFERENCE enum for DwmSetWindowAttribute's third parameter, which tells the function
-				// what value of the enum to set.
-				// Copied from dwmapi.h
-				public enum DWM_WINDOW_CORNER_PREFERENCE
-				{
-								DWMWCP_DEFAULT = 0,
-								DWMWCP_DONOTROUND = 1,
-								DWMWCP_ROUND = 2,
-								DWMWCP_ROUNDSMALL = 3
-				}
 
 				/// <summary>
 				/// Логика взаимодействия для MainDisplayView.xaml
 				/// </summary>
 				public partial class MainDisplayView : Window
     {
-								// Import dwmapi.dll and define DwmSetWindowAttribute in C# corresponding to the native function.
-								[DllImport("dwmapi.dll", CharSet = CharSet.Unicode, PreserveSig = false)]
-								internal static extern void DwmSetWindowAttribute(IntPtr hwnd,
-																																																									DWMWINDOWATTRIBUTE attribute,
-																																																									ref DWM_WINDOW_CORNER_PREFERENCE pvAttribute,
-																																																									uint cbAttribute);
         public MainDisplayView()
         {
             InitializeComponent();
-												IntPtr hWnd = new WindowInteropHelper(GetWindow(this)).EnsureHandle();
-												var attribute = DWMWINDOWATTRIBUTE.DWMWA_WINDOW_CORNER_PREFERENCE;
-												var preference = DWM_WINDOW_CORNER_PREFERENCE.DWMWCP_ROUND;
-												DwmSetWindowAttribute(hWnd, attribute, ref preference, sizeof(uint));
+												WindowSettings.SetWindowCornerRound(this);
 								}
     }
 }
